imp: Apply a transform to rotate the iFrame
Previously, the y rotation of the mesh was set to 180°/Math.PI on every
frame. This caused any updates to the y rotation to be ignored. Now, the iFrame is only rotated by 180° by applying an additional matrix3D transform to the
objects.
Also, a new file was added to show the right handed system that follows
the left handed systems file. Only thing missing is the screen rotated
by every angle.